Added second player and other fixes
Added a second player to move around with separate keys. Added key to end turn (X for Player 1, M for Player 2). Player 1 moves around with arrow keys, and Player 2 moves around with WASD.

Gravity has an issue where it works but the tanks fall down slowly, or it doesn't work at all. This needs to be fixed.