Shop
Shop implemented.

The shop will shop up at the end of the round. Players can buy things with their currency but only 1 item of each (unless there is a bug). A sound will play when purchasing, changing between players, and when a player has already purchased an item or has insufficient funds.